refactor(BA-3974): move Tus*Error to valkey_tus/exceptions.py (sibling pattern)
Match the convention used by common/health_checker, common/leader,
common/jwt, and common/message_queue/redis_queue: each sub-package
keeps its own exceptions.py next to the implementation, rather than
piling onto common/exception.py.
