+++ src/ai/backend/common/clients/valkey_client/valkey_tus/__init__.py
@@ -1,10 +1,9 @@
-from ai.backend.common.exception import (
+from .client import TusSessionId, ValkeyTusClient
+from .exceptions import (
     TusLeaseHeldError,
     TusLeaseLostError,
     TusSessionNotFoundError,
 )
-
-from .client import TusSessionId, ValkeyTusClient
 
 __all__ = [
     "TusLeaseHeldError",
